add packages for firebase_remote_config
firebase related packages set to `any` to adjust dependencies flexibly.With the newest version of firebase_remote_config, other firebase related packages are updated, which requires the higher version of kotlin. the expression of `null` is not recommended, it seems, as this indicates that the version is going to be specified but not yet.